requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
Flask==1.1.2
Jinja2==2.11.3
mysql-connector-python==8.0.24
protobuf==3.15.8
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
twilio==6.57.0
gunicorn==20.1.0
